+++ CHANGES.md
@@ -1,3 +1,7 @@
+TBD
+---
+ - Renamed to ticket-toolbox.
+
 0.1 (7 Oct 2022)
 ----------------
 Initial release.